Force split tunnel driver to remain in engaged state while blocking
